Provide product
This is required by zypp for product support

Expected is a 1-1 releation between a product (.prod file in /etc/products.d/)
and it's release package (actually the package containing the .prod file).

As the .prod file is included in a package, the repo metadata need to contain
an abstract of the product metadata (at lest everything required to properly
resolve the product).

These data are shipped in eiter repodata/products.xml
(rpmmd repo format) or in the content file (suse repo).

For each product version in the repo, its coresponding -release package is
found by looking for the package providing:

  Provides: product()
  Provides: product(<PRODUCTNAME>) = <PRODUCTVERSION>-<PRODUCTRELEASE>

<PRODUCTNAME>, <PRODUCTVERSION> and <PRODUCTRELEASE> in the package provides
and in the repos product metadata must match the products data stored in
the .prod file included in the release package.

Without product metadata or the linking Provides: the resolver does not see
that the product is continued when the -release package is updated.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
Change-Id: Iccdc3d064dab3c3312a9a8daf6c09f7285c193c0